requirements.txt
kivy[base]>=2.3.0
kivymd>=1.1.1
picamera2>=0.3.7
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
qrcode[pil]>=7.4.2
pyserial>=3.5
opencv-python-headless>=4.8.0  # For image processing
numpy>=1.24.0  # For array operations